requirements-tests.txt
tox
nose
django-nose
flake8
